Fix : recoil 사용을 위해 app 컴포넌트 생성하여 index에 렌더
